+++ rollup.config.js
@@ -6,6 +6,7 @@ import scss from 'rollup-plugin-scss';
 import visualizer from 'rollup-plugin-visualizer';
 import {nodeResolve} from '@rollup/plugin-node-resolve';
 import {terser} from 'rollup-plugin-terser';
+import json from '@rollup/plugin-json';
 
 const moduleName = 'webex-components';
 const ESModulePath = `dist/es/${moduleName}`;
@@ -13,6 +14,8 @@ const UMDModulePath = `dist/umd/${moduleName}`;
 
 const plugins = [
   nodeResolve({
+    browser: true,
+    preferBuiltins: false,
     extensions: [
       '.mjs',
       '.js',
@@ -21,6 +24,7 @@ const plugins = [
     ],
   }),
   commonjs(),
+  json(),
   babel({
     babelHelpers: 'runtime',
     babelrc: true,