+++ src/components/WebexActivity/WebexActivity.jsx
@@ -1,10 +1,12 @@
-import React from 'react';
+import React, {useContext} from 'react';
 import PropTypes from 'prop-types';
 
 import webexComponentClasses from '../helpers';
 import {useActivity} from '../hooks';
+import {AdapterContext} from '../hooks/contexts';
 
 import ActivityHeader from './ActivityHeader';
+import WebexAdaptiveCard from '../WebexAdaptiveCard/WebexAdaptiveCard';
 
 /**
  * WebexActivity component displays activity content.
@@ -16,19 +18,21 @@ import ActivityHeader from './ActivityHeader';
  * @returns {object} JSX of the component
  */
 export default function WebexActivity({activityID, className, style}) {
-  const {
-    ID,
-    created,
-    displayHeader,
-    personID,
-    text,
-  } = useActivity(activityID);
+  const activity = useActivity(activityID);
+  const adapter = useContext(AdapterContext);
+  const hasCard = adapter.activitiesAdapter.hasAdaptiveCard(activity);
+
   const [cssClasses, sc] = webexComponentClasses('activity', className);
 
   return (
-    <div className={cssClasses} key={ID} style={style}>
-      {displayHeader && <ActivityHeader personID={personID} timestamp={created} />}
-      <div className={sc('content')}>{text}</div>
+    <div className={cssClasses} key={activity.ID} style={style}>
+      {activity.displayHeader && (
+        <ActivityHeader personID={activity.personID} timestamp={activity.created} />
+      )}
+      <div className={sc('content')}>
+        {activity.text}
+        {hasCard && <WebexAdaptiveCard activityID={activity.ID} />}
+      </div>
     </div>
   );
 }